Move is_valid_token_for_ayr to app/main/authorize/keycloak_login_required_decorator.py
